requirements.txt
appnope==0.1.0
bleach==2.1.2
boto3==1.5.36
botocore==1.8.50
certifi==2018.1.18
chardet==3.0.4
colorlog==3.1.2
decorator==4.2.1
Django==2.0.2
django-enumfields==0.9.0
django-mysql==2.2.0
django-s3-storage==0.12.4
django-storages==1.6.5
docutils==0.14
entrypoints==0.2.3
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
isodate==0.6.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
mistune==0.8.3
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
oauthlib==2.0.6
olefile==0.44
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
